Store NavigationContext in CRWWKNavigationStates.
This CL (1 of 7) is a part of large refactoring to create
NavigationContext when navigation was started, not when it is finished.
This refactoring will allow to pass the same instance of
NavigationContext to DidStartNavigation and DidFinishNavigation
(crbug.com/713836). Also this will allow to associate NavigationItem
with NavigationContext object to commit non-latest navigation
(crbug.com/712269).

This CL adds API to CRWWKNavigationStates to allow associating
NavigationContextImpl with NKNavigation objects. This will allow to understand
which NavigationContextImpl should be used in WKNavigationDelegate callbacks.

BUG=713836,712269

Review-Url: https://codereview.chromium.org/2835463002
Cr-Original-Commit-Position: refs/heads/master@{#470068}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ef672aa5ffed4146ede2c1d22a04fb91dd5f168