SadTabView changes for Feedback mode
These changes represent an update to the SadTabView and its helper that uses a
simple heuristic for detecting a repeated WKWebView crash. On a repeated crash,
the UI will be altered to show a list of tips and an action to provide feedback
(rather than reload).

BUG=637410

Review-Url: https://codereview.chromium.org/2810873013
Cr-Original-Commit-Position: refs/heads/master@{#470070}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 91ac38b401efb93cec6d465c127ef7b72b421894