Return WKNavigation from loadRequest and loadPostRequest.
This CL (2 of 7) is a part of large refactoring to create
NavigationContext when navigation was started, not when it is finished.
This refactoring will allow to pass the same instance of
NavigationContext to DidStartNavigation and DidFinishNavigation
(crbug.com/713836). Also this will allow to associate NavigationItem
with NavigationContext object to commit non-latest navigation
(crbug.com/712269).

This CL makes load*Request API return WKNavigation object, which can be
used to store NavigationContext in CRWWKNavigationStates.

BUG=713836, 712269

Review-Url: https://codereview.chromium.org/2840473003
Cr-Original-Commit-Position: refs/heads/master@{#470140}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e92530274bb519ab0480a83f877ef62b8e681897