Allow storing null navigations in CRWWKNavigationStates.
This CL (6 of 7) is a part of large refactoring to create
NavigationContext when navigation was started, not when it is finished.
This refactoring will allow to pass the same instance of
NavigationContext to DidStartNavigation and DidFinishNavigation
(crbug.com/713836). Also this will allow to associate NavigationItem
with NavigationContext object to commit non-latest navigation
(crbug.com/712269).

WKWebView passes null WKNavigation to WKNavigationDelegate callbacks for
window opening navigation actions. This is a normal flow so
CRWWKNavigationStates should be able to store null navigations. Because
there can be at most one window opening navigation it is safe to use
unique key for storing null navigation.

BUG=713836,712269

Review-Url: https://codereview.chromium.org/2838593002
Cr-Original-Commit-Position: refs/heads/master@{#470439}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2c79b906de7b97f04ea867398b884ec9c0341839