Create NavigationContext right after the navigation was started.
This CL (7 of 7) is a part of large refactoring to create
NavigationContext when navigation was started, not when it is finished.
This refactoring will allow to pass the same instance of
NavigationContext to DidStartNavigation and DidFinishNavigation
(crbug.com/713836). Also this will allow to associate NavigationItem
with NavigationContext object to commit non-latest navigation
(crbug.com/712269).

This CL changes the time of NavigationContext creation. Subsequent CLs
will fix:
 - crbug.com/713836 by passing NavigationContext to
      OnProvisionalNavigationStarted
 - crbug.com/712269 by looking up for pending NavigationContext and
      extracting NavigationItem from it

BUG=713836, 712269

Review-Url: https://codereview.chromium.org/2842443002
Cr-Original-Commit-Position: refs/heads/master@{#470687}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 756b09442132dec9b01cbeda19d0a8d69e8dd551