Fix dependencies on OutOfProcessPatcher.
This is a mechanical change.

The component updater classes have dependencies on OutOfProcessPatcher.
Before this change, a client that needed to use the ComponentUnpacker
had to include component_patcher_operation.h.

With this change, the clients of ComponentUnpacker only need to include
component_unpacker.h

BUG=719751

Review-Url: https://codereview.chromium.org/2873513002
Cr-Original-Commit-Position: refs/heads/master@{#471002}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cca1c125366c60bc328af9697458aedd6c67aa17