[ios] Add session saving support to clean skeleton.
BrowserListSessionService is a KeyedService that tracks BrowserList
and automatically save the session when the current WebState of any
WebStateList changes or when a navigation item is committed in an
active WebState.

Use BrowserListSessionService to load the session during the
initialisation of the RootCoordinator.

BUG=710848

Review-Url: https://codereview.chromium.org/2858283002
Cr-Original-Commit-Position: refs/heads/master@{#471802}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b95d16dd12e99c5f1c3ea2e7316fef8f50066f98