[ios] Fix clean shutdown of Chrome on iOS.
Drop the references to SpotlightManager and BrowserViewWrangler in
MainController -stopChromeMain before destroying IOSChromeMain as
IOSChromeMain leads to the destruction of ChromeBrowserState and
both SpotlightManager and BrowserViewWrangler must not outlive the
ChromeBrowserState.

Fix SpotlightTopSitesBridge destructor by keeping a reference to
history::TopSites as __weak pointer are set to nil before the ivar
are destroyed, thus owner_->topSites is null by that point.

Fix ToolsMenuButtonObserverBridge to destroy the observer bridge
when setting _model to null to avoid DCHECK.

BUG=722879

Review-Url: https://codereview.chromium.org/2883323002
Cr-Original-Commit-Position: refs/heads/master@{#472437}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8ac159105b78c8987a16faa4113d18b37682566b