Copy the cert's refptr to the stack to avoid use-after-free crasher.
The previous code directly used the reference passed as an argument.
However, ObjC blocks treats C++ references as raw pointers, and do not
extend the scope of objects referenced within.  Copying the refptr to
the stack correctly extends the lifetime of the object.

BUG=715292

Review-Url: https://codereview.chromium.org/2894093003
Cr-Original-Commit-Position: refs/heads/master@{#473248}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ace23bbd22c0bf5523755fb7d611d6a144a8dab0