[ios] Remove TabModel cleanup from -dealloc to -browserStateDestroyed.
As the -dealloc is called at a later time with ARC, move the cleanup
from -dealloc to -browserStateDestroyed to ensure it is determinically
called even if ARC decide to -autorelease TabModel.

Cleanup some of Tab ivar when the WebState is destroyed so that the
corresponding Tab helpers -dealloc is called sooner.

BUG=None

Review-Url: https://codereview.chromium.org/2885803002
Cr-Original-Commit-Position: refs/heads/master@{#473892}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0531535acb8768a164a255cf752fce224df75ddb