[Bookmark suggestions] Remove an obsolete pref entry
This CL cleans-up an obsolete pref entry in BookmarkSuggestionsProvider.
As a result, PrefService is not needed anymore in this provider.

As a slightly unrelated cleanup, this CL removes anything bookmarks-
related from the iOS factory which does not create the provider any
more.

BUG=none

Review-Url: https://codereview.chromium.org/2895943002
Cr-Original-Commit-Position: refs/heads/master@{#473896}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c84859811406738d0392874930908f0040ecaffa