[ios clean] Creates ToolsMenu model.
Creates a new Model since re-using the old one would require too much work and
not a lot of gain. (Will expand this in the Design Doc, but Marq agreed with me that it
wasn't worth it).

This follows the pattern of the old Model but simplifies it since we won't have as many
different cases as before.

Most items are hidden, and some have been left visible just to show how it works.

TabSwitcher Menu Screenshot:
https://drive.google.com/open?id=0Byo6-Nuda2jgRjJvNnJJYkhKN1U

Web Menu Screenshot:
https://drive.google.com/open?id=0Byo6-Nuda2jgSENLR1J2T3RxRlk

BUG=682880

Review-Url: https://codereview.chromium.org/2889483002
Cr-Original-Commit-Position: refs/heads/master@{#474131}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 87a692a678babe78f8b2ddde3cf6bcc276aea8d4