Clean up the unpack directory for CRLSet component.
This is the iOS fix corresponding to:
https://codereview.chromium.org/2906683002

In the current implementation of the component updater, the component
installer owns the unpack path of a component and it is responsible
for deleting the unpack path before the execution flow returns to
the component updater during the update of a component.

For most components, this clean up is done in the implementation of
the DefaultComponentInstaller class. However, CRLSet is one of the few
components which is not reusing the DefaultComponentInstaller for
its component installer. As a result, it is leaking the content of
the unpack path after each update cycle.

It is not clear when this regression has been introduced.

BUG=725671

Change-Id: Iba01e7d9210fea4da92718cff833ea1f758935bf
Reviewed-on: https://chromium-review.googlesource.com/517582
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#475567}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 23edf9fb4b428696ecf6899e3e8e72d149f9eea5