Move common functionality of IOSChromeIOThread to ios/components.
Most of this functionality is not reliant on ios/chrome. Moving it to
ios/components allows it to be used by ios/web_view. The existing
unittest is left in place because it relies on the setup that happens in
the IOSChromeUnitTestSuite class.

BUG=725524

Review-Url: https://codereview.chromium.org/2908613002
Cr-Original-Commit-Position: refs/heads/master@{#475627}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5e8fc527dd73f46682283810a01445672e475b80