Modify visible url tests for embedded test server.
The ETS can not respond to a new request before the previous one is completed.
These two tests verify receiving of new requests while the server is paused, which
fails under ETS.  The new requests will be handled after server starts again.
Removing the checks do not change the purpose of the tests.  We can still verify the
new requests after the server is un-paused.

Modify the tests and make it run under GCDWebServer on bots first in case we are
missing some edge cases.

BUG=724560

Review-Url: https://codereview.chromium.org/2915893002
Cr-Original-Commit-Position: refs/heads/master@{#476064}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2c9727c3e7c09b51e7590f2cc429580f4a91f968