Split the ReadingListUI target in two
This CL splits the Reading List target in a UI target and a coordinator
target.
For now the ViewController are part of the coordinator target to have a clean
UI target. Once their are refactored, they will be part of the UI target.

BUG=721758

Change-Id: Id201753fa6f8bc8cc673039d940c57cc4d16d068
Reviewed-on: https://chromium-review.googlesource.com/527438
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#478247}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f4340ea66a6dee14b9bdbdccacf735aa1df428e7