[ios] Remove use of blocking pool from SnapshotCache.
Add a SnapshotCacheFactory to tie the SnapshotCache lifetime to that
of ChromeBrowserState and change client code to access it via the
factory.

Use a local base::SequencedTaskRunner to post task in background
instead of using the blocking pool. As the base::SequencedTaskRunner
keep a reference to base::TaskScheduler, the prevents SnapshotCache
from being a singleton (as it would outlive base::TaskScheduler).

Use a SEQUENCE_CHECKER to ensure the calls are made on the correct
sequence (instead on checking the current thread).

Fix unit tests to wait for TaskScheduler to become idle instead of
the blocking pool.

BUG=727597,667892

Change-Id: Icde812f74985698c4416a4dbc9c3f8b36c1e9522
Reviewed-on: https://chromium-review.googlesource.com/528193
Reviewed-by: Jean-François Geyelin <jif@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#477989}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 528c17b6de75d7aef7d7ee3e3d2abe68bb2884fe