[Autofill] Fix bug where toggling Google Payments switch was crashing cr.
Toggling the Google Payments switch toggles the
autofill::prefs::kAutofillWalletImportEnabled which is watched by the
autofill::PersonalDataManager. If as a result of this, there is any change
to the list of autofill profiles and credit cards, PersonalDataManager
notifies its observers via -onPersonalDataChanged where
AutofillCollectionViewController reloads the model. Therefore no item or
section needs to be added/removed when the swtich is toggled.

BUG=671552

Review-Url: https://codereview.chromium.org/2934453004
Cr-Original-Commit-Position: refs/heads/master@{#478566}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0d54f827187467f83c0983b6745f93fe78949e18