Showcase EG tests reset the application
When a new Earl Grey is launched in Showcase, the UI is reset with a new
collection. The close helper tests that it is returned to the home screen.
This allows the test to fail if it ends in a bad state (alert displayed...).
Before, if the test ended in a bad state, it would succeeded and the following
test would fail, making it hard to debug.

BUG=none

Change-Id: I4e65fb527fd25724c140f42193ab0ce53feeb1bd
Reviewed-on: https://chromium-review.googlesource.com/529105
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#478603}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6de61d1c4f09c354f0cad826119b21e7b92957a