Move some of IOThread's command line logic to network_session_configurator.
This will allow it to be shared with the network service, and makes it
easy to pass all relevant arguments to the network service as well.

May add handling for command line arguments that affect more than just
the HttpNetworkSession::Params there as well in a future CL, to enable
further code sharing.

TBR=lcwu@chromium.org,rkc@chromium.org
BUG=727812

Review-Url: https://codereview.chromium.org/2914893002
Cr-Original-Commit-Position: refs/heads/master@{#478740}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68a502104896fc2bb14fc3903afe31d275e2ffda