Refactoring bookmark viewcontrollers
BookmarkHomeHandsetViewController and BookmarkHomeTabletNTPController
have a lot of overlap in terms of business logic, and hence results
in maintenance overhead. This effort is to refactor the classes so that
they have a common super class with the core logic.

This is CL#1: Merge existing BookmarkHomeViewController and
BookmarkHomeHandsetViewController into one class.

See go/bookmarks-refactor for more details.

BUG=705339

Review-Url: https://codereview.chromium.org/2929993002
Cr-Original-Commit-Position: refs/heads/master@{#478918}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 742eb0b40de63b805481c7816f4de124d46211af