Remove workaround for strikethrough text in iOS omnibox
iOS 10.3 introduced a regression in rendering of strings with the
attribute NSStrikethroughStyleAttributeName applied to a subrange of
the string. For crbug.com/699702, we introduced a workaround that is
no longer needed as of iOS 10.3.2. We will remove that workaround now
that iOS 10.3.2 has released.

BUG=711367

TEST=https://expired.badssl.com. Examine https: text in the omnibox,
it should have a line through it.

Review-Url: https://codereview.chromium.org/2934473003
Cr-Original-Commit-Position: refs/heads/master@{#479799}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 927047b0999e1fa6c164d581bdd5ec3b0e4d1f64