Turn on services_unittests on the iOS bots
These tests are running on the bots on every other platform, so let's
turn them on on iOS as well. This should not be a problematic rollout
as services_unittests on iOS is a quite minimal set of tests at the
moment. However, if there are problems, feel free to revert this CL.

Bug: 731582
Change-Id: I612d39c0ff52d6a3ccdfd3e22a80041aa682d901
Reviewed-on: https://chromium-review.googlesource.com/538676
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#480376}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6a0eec4e4305d89c30068ef402f3de6f1c578cf1