Remove custom resizing hooks in GoogleLandingVC
GoogleLandingViewController used to have custom hooks to detect resizing events
or orientation changes. This CL removes them and only use the generic
UIViewController hooks.
It also change the NTP constraints for the logo and the omnibox: the constraint
and positioning is the same in landscape and portrait. The omnibox and doodle
are now higher in portrait.

TEST= This CL refactors the NTP. Some regression testing should be done, in particular on rotations.
Issues for which specific fixes have been removed: crbug.com/491131
On iPad and in fullscreen, the collection view's inset is very large. When Chrome enters slide over mode,
the previously set inset is larger than the newly set collection view's width, which makes the collection view throw an exception.

Bug: 733166
Change-Id: I2f6faf2468196cdd0a2faa191634b4da51254079
Reviewed-on: https://chromium-review.googlesource.com/536915
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#480453}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f6e429ff2786b7d1d2c16bd4902fa4b4b9ee5ce5