Return nil if decoded NSData cannot be converted to a X509Certificate.
In CRWSessionCertificateStorage's previous NSCoding implementation, it
was possible for a cert storage to be created where its |certificate|
property returns nil.  But CL updates its implementation so that if the
certificate can't be decoded properly, the object will not be
instantiated at all.

BUG=731117

Review-Url: https://codereview.chromium.org/2940843004
Cr-Original-Commit-Position: refs/heads/master@{#480700}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e45a8bf9aadce889de072a407eea573f73ce52b0