Add NetLog events for HttpServerPropertiesManager
This CL adds NetLog events for HttpServerPropertiesManager.
(1) HTTP_SERVER_PROPERTIES_INITIALIZATION
This is emitted when initialization starts and finishes. This is useful to know
if server information is used in making network requests.
(2) HTTP_SERVER_PROPERTIES_UPDATE_CACHE
This is emitted when we are populating the in-memory HttpServerProperties from
the on-disk version.
(2) HTTP_SERVER_PROPERTIES_UPDATE_PREFS
This is emitted when we are persisting the in-memory HttpServerProperties to
disk.

A sample NetLog:

https://drive.google.com/a/chromium.org/file/d/0Bw0C8-TgzJfsVHVmaTJacjNlQzg/view?usp=sharing

BUG=718576
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.android:android_cronet_tester

Review-Url: https://codereview.chromium.org/2944043002
Cr-Original-Commit-Position: refs/heads/master@{#481267}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb70779863bf9dfe419601a6de536c336feaf591