Gracefully handle didCommitNavigation: call after didFinishNavigation:
It is possible to get WKNavigationDelegate callbacks in the following
order: didFinishNavigation, didCommitNavigation:.

In this case didFinishNavigation should not remove navigation from
CRWWKNavigationStates storage, because that navigation will be used in
didCommitNavigation.

Also didCommitNavigation should remove navigation and signal
didFinishNavigation: which is expected to be called only after
navigation is committed.

BUG=727289

Review-Url: https://codereview.chromium.org/2950853002
Cr-Original-Commit-Position: refs/heads/master@{#481584}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5c9be243c093b5410523730bdd377dd4cd68578d