Removed XCallbackParameters and related code.
XCallbackParameters does not carry anything information that is used
and is just passed around the app startup parameters due to legacy
reasons.

Note that googlechrome-x-callback and chromium-x-callback custom URL
schemes are *still supported*. They are parsed and the URL part is
extracted for opening in a tab.

Review-Url: https://codereview.chromium.org/2923013005
Cr-Original-Commit-Position: refs/heads/master@{#481885}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 489d28390c0aee4c59b9c2621bb67930b579d007