[iOS] Set credit card icon using the saved network.
When possible, the code stops using the credit card number to obtain the network.
This is because the browser only knows the last four digit or non local cards, which is
not sufficient to know the network.

This CL also changes the label shown for non local credit cards:
Instead of containing "1234", the label contains "Mastercard ... 1234".

BUG=719988

Review-Url: https://codereview.chromium.org/2949283002
Cr-Original-Commit-Position: refs/heads/master@{#482265}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4e19586ba1c5e3e5c837594366b89cb936928fc9