Possibility to add exceptions in Protocol Alerter
The ProtocolAlerter logs all the methods of a protocol. In some cases,
a showcase example does not want to log all methods, because their are
called too often or are not relevant.
This CL adds a way to remove the logging for some specific methods.

BUG=none

Change-Id: I7956911e6f472271a57e72288c7ee04d86623084
Reviewed-on: https://chromium-review.googlesource.com/549898
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#482610}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 47642f05bc6abb99b8df5f3900e5c89a5e88ef9a