[ios] TextBadgeView interface, implementation, and unit test stubs.
Added header, implementation, and unit test files for TextBadgeView.
Updated the BUILD.gn file in the reading_list directory to include the
header and implementation files in the reading_list_ui source set, and
the unit test file in the unit_tests source set. The header,
implementation, and unit test files are stub implementations.

For context, the TextBadgeView is part of the User Education project.

BUG=none.

Review-Url: https://codereview.chromium.org/2954853004
Cr-Original-Commit-Position: refs/heads/master@{#483205}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e28c24b9848609678128c3644f752bcf8d5832d7