Fix the NTP shadow
The bottom toolbar shadow was present when the NTP opened on iPhone, then
disappeared if it was not needed, based the height of the device.
This CL changes it to have it adjusted to the device height from the beginning.

BUG=736057

Change-Id: I48d41a1b6230d622a4872b0d0c34959516218729
Reviewed-on: https://chromium-review.googlesource.com/549358
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#483371}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 26777237dc6c12f18afa1304f05647de0506b3b7