Move tiles up/down when (un)focusing the omnibox
This CL rewrites the way the HeaderController and the CollectionController
are communicating with each others.
The ContentSuggestionsHeaderCollectionCommandHandler is handling the
interactions by conforming to the two commands protocols.
It handles the move of the tiles and the updates of the fake omnibox.

BUG=700375

Change-Id: Id465531d7c09b3713c184307a247e0db897e5bcf
Reviewed-on: https://chromium-review.googlesource.com/552148
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#483987}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 202930499e5efc7bddb88f4a382a1b6be1d61d02