Switch AffiliationService from DB thread to a TaskRunner
This CL does these significant changes:
(1) Changes a SingleThreadTaskRunner to SequencedThreadTaskRunner in
    AffiliationService and related backend code.
(2) Pushes obtaining of the TaskRunner to the callsite of the
    AffiliationService constructor.
(3) Replaces the choice of BrowserThread::DB thread task runner with a generic
    one which has the lowest priority and sequential execution guarantee.

I could not find a reason why single thread (thread affinity) was required by
the affiliation code, so my hope is that (1) is safe. The tasks to be posted
on the task runner are potentially blocking and are not part of user
interaction, so setting the BLOCKING trait (currently the lowest priority)
seems appropriate, explaining (3).

As for (2), the general guidance [1] is that task runner is obtained down in
the component by the "end user" of it. This is not easily done, however,
because the unittests for AffiliationService inject a different task runner,
which enables time-forwarding simulation for testing. Still, at least now
the task runner is obtained in the component.

[1]
https://chromium.googlesource.com/chromium/src/+/master/docs/task_scheduler_migration.md#BrowserThreads,
look for "passed down"

R=engedy@chromium.org
BUG=689520

Review-Url: https://codereview.chromium.org/2968903002
Cr-Original-Commit-Position: refs/heads/master@{#484017}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4cb77ce419b39c1fd06975f16374b6cee0f06224