Adding events in ChromeBrowserProvider
Adding:
  OnChromeIdentityServiceDidChange()
  OnChromeBrowserProviderWillBeDestroyed()

In the UI tests, ChromeIdentityService is changed between test case, but the UI
is not reconstructed. The sign-in promo mediator has to listen events from the
ChromeIdentityService. So to be able to create UI tests for the sign-in promo
view, events about changing ChromeIdentityService has to be created so the
mediator can pickup the new ChromeIdentityService.

BUG=661794

Review-Url: https://codereview.chromium.org/2969903002
Cr-Original-Commit-Position: refs/heads/master@{#484019}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7c72ce06f5f925b7309ede559ba7d8880ea1146c