Fix search widget UI so animation is not choppy.
How this works: The actions (top row of circle buttons) and copied URL
views are always visible. The actions view's height is set to the
full widget height when the widget is in compact mode. This pushes the
copied URL view out of the screen. By not changing relying on constraints
to get the widget preferred size, the change can be made at the right
time to allow animation to work correctly. See video below for an
example of the widget animating.
https://drive.google.com/a/chromium.org/file/d/0BxrfNSu8yqJqcXhMQmJqeWNVQ00/view?usp=sharing

Bug: 622743
Change-Id: I02f2bb091da2129ea025c878b6a75f15ec984a53
Reviewed-on: https://chromium-review.googlesource.com/559052
Commit-Queue: Elodie Banel <lod@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#484101}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36b73fc4f31d8054f05c4bf02eecf0f36e7011c0