Add ContentSuggestionsFaviconMediator
This CL adds a ContentSuggestionsFaviconMediator, charged to fetch the
favicons for the ContentSuggestions items. This object lives in the
mediator layer.
This CL also modifies the fetching to initiate it in the mediator
instead of the ViewController.

BUG=736002

Change-Id: I10e25462f38145354c511474e390640da7e01b9a
Reviewed-on: https://chromium-review.googlesource.com/558997
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#484103}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a52bd263e1713082d7048dacd1964448ec4df0e3