Fix a raw ptr vs. unique_ptr clash in password settings test
SavePasswordsCollectionViewController has an internal method
onGetPasswordStoreResults which takes a vector of unique pointers.

Its unittest declares that method to gain access to it, but with a different
type: taking a vector of raw pointers instead. It puts heap-allocated objects
to those pointers and never frees them.

Astonishingly, this setup works. I suppose it is because unique_ptr stores the
pointer at offset 0, and Objective C does not somehow care about types as
strongly as C++, so when the production code reinterpreted the raw pointers as
unique_ptr, it worked and the objects got even freed?

Anyway, that seems to be working just by accident, so this CL unifies the
types to be unique_ptr on both sides.

R=vasilii@chromium.org

Bug: 
Change-Id: I77d28709e264d8b453161b77ccb34de12fc27eff
Reviewed-on: https://chromium-review.googlesource.com/559125
Commit-Queue: Vaclav Brozek <vabr@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#484248}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e470f1da1e847d7c0b80bffd9632b785253123b5