ToolbarModel: Supply offline page status.
This allows users of ToolbarModel to determine whether the current page
is an offline page or not.  In the future, it'd be interesting if the
security level, vector icon and verbose security text could simply
indicate that the page is offline, rather than using a separate bit of
information.

BUG=735770

Review-Url: https://codereview.chromium.org/2965043003
Cr-Original-Commit-Position: refs/heads/master@{#484789}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d0af1dee22043449ff2076b1ba361a47b1803b3d