[ios] Remove IDC_STOP and IDC_RELOAD
This CL replaces the IDC_STOP and IDC_RELOAD command-ids with selectors
in BrowserCommands. Call sites and ancillary usage is also updated.

This involves:

- Plumbing a dispatcher into the PageInfoController and having the 
reload button in that class target the dispatcher with the right method.

- Plumbing a dispatcher into the PopupMenuController (and subclasses), 
and the ToolsMenuViewController.

This CL also establishes the pattern for migrating tools menu items to
use the dispatcher. It is unfortunately hacky. The tools menu uses
the existing command IDs for two purposes: to tag the menu item so it
can be sent with chromeExecuteCommand, and for -commandWasSelected to
key off of to increment the right metric. Removing the first use (the
primary goal of this CL) means finding something else to use for the
second, and in this case it's a new enum defined in tools_menu_constants.h

These enum values are arbitrary *except* they are unique and need to be
distinguished from command IDs; to fit the latter criterion they are all
negative values.

This all comes together in ToolsMenuViewController:

- -collectionView:cellForItemAtIndexPath:, which already attaches
  -buttonPressed: as an action to every button in the menu, also 
  attaches the dispatcher methods for the stop and reload buttons.
- hideContent, similarly, removes those as actions.
- -buttonPressed: used to send ChromeExecuteCommand; now it only does
  so if the commandID is > 0. I realize this is terrible.

Because -buttonPressed, via two further levels of delegation(!), is 
responsible for dismissing the menu itself, the dispatcher methods must
be called *before* buttonPressed:, and so are added as actions before that
method is.

Note that because some downstream code uses IDC_RELOAD, the command ID
itself isn't removed in this CL, and the BVC switch still handles it.
That will be cleaned up in a follow-up CL.

Bug: 738881
Change-Id: Ibe2643c80b399e7a8ddee11c42c1ab7a569d7f2f
Reviewed-on: https://chromium-review.googlesource.com/561678
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#484909}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b9aac643aaa7a01bccc7abde501dafbad78036f2