[ios] Remove IDC_SHOW_TOOLS_MENU
This CL removes the IDC_SHOW_TOOLS_MENU command, and updates call sites.

This involves:

- Adding a command dispatcher to the stack view. This is done by defining
a dispatcher property on the TabSwitcher protocol; StackViewController
creates a dispatcher on init and returns it for this property. The tab
switcher view controller returns nil (for now).

- The special case in the tools menu on phones for the menu button to
no-op but still record metrics is removed by simply not having the 
tools menu button in that case call -showToolsMenu.

(Note that tapping the tools menu button while the menu is showing is
still a no-op, presumably because that's easier than special-casing
the button to do nothing. Since all of the menu and toolbar code is 
moribund, this doesn't seem worth cleaning up.)

Bug: 
Change-Id: I6d90ff72cdc6969679ec1c5e07dd6ee32cf0deed
Reviewed-on: https://chromium-review.googlesource.com/563660
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#485565}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6acee7f8db26d5fb3457b740f886fbb41eb096d4