Fixing impression count for settings.
The "ImpressionsTilDismiss" should be sent only when the setting view
is closed (it was sent when going into a submenu of the settings).
The display count should only be increased when opening the settings.
Before it was increased for each data reload of the collection view.

Bug: 709286
Change-Id: I301f96241f7704aa01695fc41e841364fca1f80a
Reviewed-on: https://chromium-review.googlesource.com/563389
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#485574}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8baa47c590dd255381432cb7db37a869a8b7c452