[ios] Remove IDC_OPTIONS
This CL removes IDC_OPTIONS. Since this is handled by MainController and
not a BVC, this CL adds the plumbing necessary for this.

Commands handled by the MainController are grouped into ApplicationCommands;
the existing dispatchers will dispatch for this protocol. The BVCs and 
tab switchers will be handed an 'Application command endpoint' that they
can dispatch that protocol to. 

This means that the various id<BrowserCommands> 'dispatcher' properties 
become id<ApplicationCommands, BrowserCommands>s. That accounts for many
of the small changes in this CL. It is clunky for formatting in some cases,
so when all of the refactoring is done, a clean-up CL might be needed that
groups protocols together for brevity.

The other major change is adding a selector parameter to the tools menu
model. Menu items that use a dispatcher have a nonnull selector. Note that
because a "command ID" is still needed for metrics, converted items get a
TOOLS_<blah>_ITEM ID as well.

Menu items with selectors use -performSelector: on their dispatcher to 
send the relevant method.

Bug: 738881
Change-Id: I1cce22e20b96f8593ca3b74f5e54b8ccf5fe6c8e
Reviewed-on: https://chromium-review.googlesource.com/565508
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#485640}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5e3da1596d592fcd2db55b4580510ecc6a2b833c