[iOS passwords settings] Tie haptic feedback to all copy toasts
Currently, in passwords settings, haptic feedback is given during copying
password but not during copying any of the other less sensitive data (site
and username).

https://crbug.com/696634 hints at, and pinkerton@ confirmed, that the idea
for haptic feedback was to accompany all MD toasts in the UI. pschaffner@
confirmed on this CL that that's a good idea for the passwords settings
as well.

Therefore, this CL adds the haptic feedback to the site and username
copy toasts. Those copies always succeed, because there is no
reauthentication to be done, so the feedback is always a successful
one.

Bug: 740948
Change-Id: Ie282499eb5095cabeb4d8349f6231ae653f01f6a
Reviewed-on: https://chromium-review.googlesource.com/567152
Commit-Queue: Vaclav Brozek <vabr@chromium.org>
Reviewed-by: Mike Pinkerton <pinkerton@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#486107}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5201a681d04dc0fefdfca0b1eca8b3df2e40c9eb