[Payment Request] Makes PaymentDetails.total property optional for updates.
In PaymentDetailsUpdate which is used to update the payment request using
updateWith(), total is an optional field. On the other hand, in
PaymentDetailsInit which is used in the construction of the payment request,
total is a required field. This CL allows for that distinction.

Bug: 602666
Change-Id: I2109a1b5af1bad881d6e3bb63590448ac16aabd0
Reviewed-on: https://chromium-review.googlesource.com/568286
Commit-Queue: mahmadi <mahmadi@chromium.org>
Reviewed-by: Mathieu Perreault <mathp@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#486170}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 38d333d060245ee5fda5c3a76254e3cf6bea1fe5