Reland "Disable prefetching on ContentSuggestions"
Prefetching allows the collection view to get cells configured by the
item before it is visible. The item will not reconfigure the cell for
this display.
If the CollectionViewController tries to reconfigure the items
associated with this cell in the moment between the prefetching and
the time the cell is visible, the reconfigure will do nothing.

In ContentSuggestions this problem was encounter for the display of
the articles' images.

Bug: 740572
Change-Id: Ia9138b22a13796d6079f04113f95dab1f30b6fb5
Reviewed-on: https://chromium-review.googlesource.com/569162
Reviewed-by: Elodie Banel <lod@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#486320}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 755f0dd1940844042792972fe80755c95f6d8b29