[Payment Request] Adds tests for accessibility on main pages of Payment Request
The added HTML file and JS file have been downloaded from here:

https://rsolomakhin.github.io/pr/multi/

And thus do not need to be thoroughly reviewed. The only changes made to these
files are as follows
1) In the html file an "id" attribute is added to the "buy" button, the 2016
copyright is changed to 2017, and the stylesheet and favicon links are removed.
2) In the js file "requestPayerEmail," "requestPayerPhone," and "requestPayerName"
is added to options on line 106, the JSLint directives are removed from the top of
the file, and the long line on line 137 was changed into two lines.

Bug: 734571
Change-Id: Ib27f0e9398d3ce8922018138ad98e0e93aa6e15b
Reviewed-on: https://chromium-review.googlesource.com/567215
Commit-Queue: Randall Raymond <rayraymond@google.com>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Reviewed-by: mahmadi <mahmadi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#486391}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 944975a0b6009211490a960f54530452c73dfb96