Always set the navigation pending item for page reload.
Set the pendingItemIndex to the last committed item in case of page reload
and on the case of application open/tab out of memory reopened.

This doesn't solve the PendingItem being null problem entirely there are still
couple cases when that can happen and they are not page reloads.

CL 1/2

BUG=676129

Review-Url: https://codereview.chromium.org/2931833004
Cr-Original-Commit-Position: refs/heads/master@{#486777}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc2e0a45e23c739bb3c04d119d43abbb01398289