Don't send dismiss histogram if the user logged in
On iPad, the user can show the bookmarks with the sign-in promo and open
the settings. If they sign in with the settings, then the sign-in promo
will be removed from the bookmarks. The mediator from the bookmarks
should not send the dismiss histogram.

To avoid this problem, the sign-in promo view mediator needs to check
if the user is logged in before sending the dismiss histogram.

Bug: 742514
Change-Id: Id5c189e68a06e82874ef487155831593911a9ef4
Reviewed-on: https://chromium-review.googlesource.com/567508
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#487061}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 74765b491b95983786e51bd187bb7aa1c31a04c9