[Payment Request] Adds tests for accessibility on main pages of Payment Request - Reland
Two build bot json files have been edited so that they run with the appropriate 
command line flags. Given that these build bot changes are related to adding to this
test this is all being packaged as one CL.

From original CL description:
"The added HTML file and JS file have been downloaded from here:

https://rsolomakhin.github.io/pr/multi/

And thus do not need to be thoroughly reviewed. The only changes made to these
files are as follows
1) In the html file an "id" attribute is added to the "buy" button, the 2016
copyright is changed to 2017, and the stylesheet and favicon links are removed.
2) In the js file "requestPayerEmail," "requestPayerPhone," and "requestPayerName"
is added to options on line 106, the JSLint directives are removed from the top of
the file, and the long line on line 137 was changed into two lines."

TBR=@lpromero@google.com -- Reviewed already in last past. Only addition here is the
build bot changes.

Change-Id: Ia20efa783eb763ae97f7e0df2f80e710823d7482
Reviewed-on: https://chromium-review.googlesource.com/570879
Commit-Queue: Randall Raymond <rayraymond@google.com>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: mahmadi <mahmadi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#487162}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 535fbc0b73aafc1e7a5346c5347802c1dc3d485b