Delay binding of SourceId to URL in PasswordManagerMetricsRecorder
This CL delays the binding of a ukm::SourceID to a URL in the
PasswordManagerMetricsRecorder until the recorder is destroyed. The reason for
this is that long-lived tabs/WebContents would bind the URL at creation. But by
destruction (i.e. reporting time) the binding is purged and the data is
discarded.

Bug: 732846
Change-Id: I4d20385eb2de80c4ca5e53816d22b9652abb5b61
Reviewed-on: https://chromium-review.googlesource.com/577554
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Commit-Queue: Dominic Battré <battre@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#488192}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9dc9d281b4bd298555bcbc112e2911a3d4a67959