[iOS XCUITest] Fix XCUITest code sign errors.
The code sign script has been updated to assert that the application
binary name and the CFBundleName must be the same as the bundle name
without the .app suffix.

This CL updates the XCUITest template to comply with the rule.

Bug: 709289
Change-Id: I5e4fa8afa07e16289903b7523fe647be04351ac3
Reviewed-on: https://chromium-review.googlesource.com/576454
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#488257}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bbb45a5366b9db4aed888bf1fef99e8a731f3e10