Retain config in WKWebViewConfigurationProviderTest.Purge
On release build, the weak config is deallocated even before the
ASSERT_TRUE.
Add a strong variable to retain it at least in the autoreleasepool.

Note that this test does not test the deletion of the config in the
Purge method as GetWebViewConfiguration returns a copy.

Bug: None
Change-Id: If3e21c3bab53c35859ed1b70c3b207966a02813f
Reviewed-on: https://chromium-review.googlesource.com/579430
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#488260}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b8a82b57b13d631560f5685fc9035a0a33d4e5ae