Make showcase home screen not translucent.
EarlGrey 1.10.1 has a bug that the scrollToTop action hangs
indefinitely if the UITableView's navigation bar is translucent. Given
that it makes no difference whether the view is translucent or not,
this CL makes it not translucent to unblock rolling EarlGrey to 1.10.1.

Bug: 747116
Change-Id: Idd48cf6f3f024c4b88bedcca412692598cd1a493
Reviewed-on: https://chromium-review.googlesource.com/580357
Reviewed-by: Ed Chin <edchin@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#488417}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 17aced1485e54a55ab1a1473650c5325b2158e45